elevenlabs_module/requirements.txt
python-dotenv==1.0.1
elevenlabs<0.2.27
pydantic<2.0
loguru==0.7.2
requests>=2.32.0
sounddevice>=0.4.6
scipy>=1.11.0